Move OTEL init fixture to main fixtures module
